requirements.txt
Khayyam==3.0.17
pillow==10.4.0
